Temporarily disabled SVG config
Custom processing of SVG file has issues where web browsers are unable to render the output.

It's believed to be due to the main xmlns being necessary, but haven't been able to verify absolutely yet.  In the meantime, disabling in favour of the default pipeline until it can be fixed in the future.